fix all register bug

+++ src/utils/cloudinary.js
@@ -1,30 +1,34 @@
-import {v2 as cloudinary} from "cloudinary";
-
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv";
+
+dotenv.config ();
+
 
 cloudinary.config({
-    cloud_name:process.env.CLOUDINARY_CLOUD_NAME ,
-    api_key:process.env.CLOUDINARY_API_KEY,
-    api_secret:'process.env.CLOUDINARY_API_SECRET',   
-    })
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 
+const uploadOnCloudinary = async (localFilePath) => {
+  try {
+    if (!localFilePath) return null;
 
-    const uploadOnCloudniary = async (localFilePath) => {
-        try {
-            if (!localFilePath) return null
-            //upload file on cloudinary
-            const response  = await cloudinary.uploader.upload(localFilePath,{
-                resource_type:"auto",
-        })
-        // file has been uplaod succesfuly 
-        console.log ('file is uploaded on cloudinary', response_url);
-        return response;
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
 
-    }catch (error){
-
-        fs.unlinkSync(localFilePath);
-        // remove the locally save tempored file as the upload opration got failed 
-
+    console.log("✅ File uploaded on Cloudinary:", response.url);
+    return response;
+  } catch (error) {
+    console.error("❌ Cloudinary upload error:", error);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-}
-export {uploadOnCloudniary};
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
+
